# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogLogSource": {
        "additionalProperties": false,
        "description": "One `catalog_log_sources` row.",
        "properties": {
          "context_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Note"
          },
          "criticality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ]
          },
          "defined": {
            "title": "Defined",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "in_scope": {
            "title": "In Scope",
            "type": "boolean"
          },
          "log_source_id": {
            "title": "Log Source Id",
            "type": "integer"
          },
          "missing_since": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Since"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "type_name": {
            "title": "Type Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "log_source_id",
          "name",
          "type_name",
          "defined",
          "description",
          "owner",
          "criticality",
          "in_scope",
          "context_note",
          "updated_by",
          "updated_at"
        ],
        "title": "CatalogLogSource",
        "type": "object"
      },
      "CatalogLogSourceUpdate": {
        "additionalProperties": false,
        "description": "`PUT /catalog/log-sources/{log_source_id}` (admin). The log source becomes defined.",
        "properties": {
          "context_note": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Note"
          },
          "criticality": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "in_scope": {
            "title": "In Scope",
            "type": "boolean"
          },
          "owner": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          }
        },
        "required": [
          "in_scope"
        ],
        "title": "CatalogLogSourceUpdate",
        "type": "object"
      },
      "CatalogMode": {
        "enum": [
          "analyze",
          "skip"
        ],
        "title": "CatalogMode",
        "type": "string"
      },
...
      "CatalogRuleUpdate": {
        "additionalProperties": false,
        "description": "`PUT /catalog/rules/{rule_id}` (admin). The rule becomes defined.",
        "properties": {
          "attack_techniques": {
            "items": {
              "pattern": "^T[0-9]{4}(\\.[0-9]{3})?$",
              "type": "string"
            },
            "maxItems": 32,
            "title": "Attack Techniques",
            "type": "array"
          },
          "context_note": {
            "anyOf": [
              {
                "maxLength": 600,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Note"
          },
          "has_automated_action": {
            "title": "Has Automated Action",
            "type": "boolean"
          },
          "min_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "$ref": "#/components/schemas/CatalogMode"
          }
        },
        "required": [
          "mode",
          "has_automated_action"
        ],
        "title": "CatalogRuleUpdate",
        "type": "object"
      },
...
      "Page_CatalogLogSource_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogLogSource"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[CatalogLogSource]",
        "type": "object"
      },
      "Page_CatalogRule_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogRule"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[CatalogRule]",
        "type": "object"
      },
...
      "Problem": {
        "description": "The shape of an RFC 9457 answer, for the UI's own error handling.\n\nNot a response model of any route: `ais0c_api.openapi` makes it every operation's `default`\nresponse in the schema. `errors`, `domains` and `list_names` appear only on the problems that\ncarry them.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Domains"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "list_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "List Names"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The analyst UI's API. The UI talks only to this service.\n\nErrors are RFC 9457 problems whose `title` is a machine-readable code (`catalog.rule_not_found`,\n`qa.already_resolved`); the Turkish message the user reads is built by the UI from that code.\nTimes are ISO 8601 in UTC. Lists are cursor-paged with `?cursor=&limit=`.\n",
    "title": "ais0c analyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/catalog/log-sources": {
      "get": {
        "description": "The log sources by ID. `q` matches part of the name or the type name.",
        "operationId": "get_log_sources_api_v1_catalog_log_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "defined",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Defined"
            }
          },
          {
            "in": "query",
            "name": "in_scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "In Scope"
            }
          },
          {
            "in": "query",
            "name": "missing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Missing"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CatalogLogSource_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Log Sources",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/log-sources/{log_source_id}": {
      "get": {
        "description": "One catalog log source; an unknown ID is a 404.",
        "operationId": "get_log_source_api_v1_catalog_log_sources__log_source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "log_source_id",
            "required": true,
            "schema": {
              "title": "Log Source Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogLogSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Log Source",
        "tags": [
          "catalog"
        ]
      },
      "put": {
        "description": "An admin's edit of a log source. It becomes `defined` and the change is audited.",
        "operationId": "put_log_source_api_v1_catalog_log_sources__log_source_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "log_source_id",
            "required": true,
            "schema": {
              "title": "Log Source Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogLogSourceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogLogSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Put Log Source",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/rules": {
      "get": {
        "description": "The rules by ID. `q` matches part of the name; `missing` keeps what QRadar stopped listing\n(`missing_since` set) or only what it lists (`missing=false`) (T-37).",
        "operationId": "get_rules_api_v1_catalog_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "defined",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Defined"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CatalogMode"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "in": "query",
            "name": "qradar_enabled",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Qradar Enabled"
            }
          },
          {
            "in": "query",
            "name": "missing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Missing"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CatalogRule_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Rules",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/rules/{rule_id}": {
      "get": {
        "description": "One catalog rule; an unknown `rule_id` is a 404.",
        "operationId": "get_rule_api_v1_catalog_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogRule"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Rule",
        "tags": [
          "catalog"
        ]
      },
      "put": {
        "description": "An admin's edit of a rule. The rule becomes `defined` and the change is audited.\n\nAn unknown `rule_id` is a 404 and nothing is written.",
        "operationId": "put_rule_api_v1_catalog_rules__rule_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogRuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogRule"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Put Rule",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/rules/{rule_id}/accept-draft": {
      "post": {
        "description": "Accept the note the AI suggested: it becomes the rule's `context_note` and the draft is\ncleared. A rule with no draft, or an unknown `rule_id`, is a 404.",
        "operationId": "post_accept_draft_api_v1_catalog_rules__rule_id__accept_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogRule"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Post Accept Draft",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/catalog/sync": {
      "post": {
        "description": "Start the KnowledgeSync Schedule now; Temporal runs it, and 202 says it was accepted.\n\nThe API itself reads nothing from QRadar. The audit row is written first and committed only\nwhen Temporal took the trigger, so a sync never starts without one. Temporal being\nunreachable is a 503 (`temporal.unavailable`); a Schedule the batch worker has not created\nyet is a 409 (`catalog.sync_not_scheduled`). Neither leaves an audit row.",
        "operationId": "post_sync_api_v1_catalog_sync_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Post Sync",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}